Moves number util functions to core
