+++ src/components/number-display/number-display.tsx
@@ -1,13 +1,10 @@
-import React from "react";
+import {
+  NumberLike,
+  formatNumber as formatNumberDefault,
+  parseNumberLike,
+} from "~/core/number";
 
-/**
- * Values that can represent a number.
- *
- * NOTE: You may decide to accept strictly numbers instead and that is valid!
- * If so, you can most likely remove this type and the idea of "parsing".
- * Accepting strings can be useful though for dealing with user input, API responses, etc.
- */
-export type NumberLike = number | string;
+import React from "react";
 
 type NumberDisplayProps = {
   value: NumberLike | null | undefined;
@@ -29,38 +26,6 @@ type NumberDisplayProps = {
    */
   fallback?: JSX.Element | string | null;
 };
-
-function isValidNumber(x: unknown): x is number {
-  return typeof x === "number" && !isNaN(x);
-}
-
-/**
- * A default formatter leveraging the browser's locale.
- */
-export function formatNumberDefault(
-  d: number,
-  opts: Partial<Intl.NumberFormatOptions> = {}
-): string {
-  return d.toLocaleString(undefined, {
-    maximumFractionDigits: 2,
-    ...opts,
-  });
-}
-
-/**
- * A default parser which accepts a range of inputs.
- * It is likely different apps will have different needs and this is swappable in `NumberDisplay`.
- */
-export function parseNumberLike(
-  x: NumberLike | null | undefined
-): number | null {
-  if (typeof x === "string") {
-    const result = Number(x);
-    return isValidNumber(result) ? result : null;
-  }
-
-  return isValidNumber(x) ? x : null;
-}
 
 /**
  * By default, renders the formatted number with no wrapping markup.
