+++ src/core/number/index.ts
@@ -1,0 +1,1 @@
+export * from "~/core/number/number-core";
